requirements.txt
obs-websocket-py==0.4